+++ include/data/sa_x_close_up.h
@@ -5,9 +5,9 @@
 
 extern const u16 sSaXCloseupPal[16 * 16];
 extern const u16 sSaXCloseupPalGreyscale[16 * 16];
-extern const u8 sSaXCloseupGfx[];
-extern const u8 sSaXCloseupFirstTileTable[];
-extern const u8 sSaXCloseupSecondTileTable[];
+extern const u32 sSaXCloseupGfx[];
+extern const u32 sSaXCloseupFirstTileTable[];
+extern const u32 sSaXCloseupSecondTileTable[];
 
 extern const u8 sSaXCloseupDurations[2];
